move service_monitor out to separate file
